Modified Cockpit to list a much more limited set of files when determining whether S3 objects clash with local files. Prior to this change, the app could run out of memory if you tried to download files to a local directory that contained many files (eg 250,000+)
To effect this change, a modification was required in FileComparer#buildFileMap to ensure non-existent files are ignored when building a file map. This in turn required a change in Synchronize to check for the existence of files before generating a file map, to ensure non-existent files cause an error rather than being ignored.