AWS version 4 signing now works with payloads if SHA256 pre-set, re #183
Requests with a payload are now passed through the signature signing
process correctly, provided the SHA256 hash value is pre-set on the
object as the "x-amz-content-sha256" header.

The S3Object helper constructors that build an object from a string or a
file automatically calculate and set this value, so objects constructed
the default way will just work.

Also unit-tested URI sigining for non-latin characters, which showed
some URI encoding problems which are now fixed.