Add AWSEC2IAMSessionCredentials load method to auto-detect role name, re #163
If a single, unique and unambiguous role name value is available at the EC2
instance data endpoint [1], a new variant of the static data-loading method
AWSEC2IAMSessionCredentials.loadFromEC2InstanceData will detect and use that
role name.

Thanks to Felix Natter (fnatter) for the contribution.

Disclaimer:

I find the relevant documentation too confusing to know whether or when
a single role name is available and safe to use in this way -- but if it works
in practice it will save hard-coding or configuring another unnecessary piece
of data.

[1]: http://169.254.169.254/latest/meta-data/iam/security-credentials/